+++ youtubedl/templates/part.settings.php
@@ -1,12 +1,18 @@
 <div id="app-settings">
 	<div id="app-settings-header">
-		<button class="settings-button"
-				data-apps-slide-toggle="#app-settings-content"
-		></button>
+		<button class="settings-button">Settings</button>
 	</div>
-	<div id="app-settings-content">
+	<div id="app-settings-content" style="display:block;">
         <ul>
-            <li><a href="?refreshCache=true">Refresh Folder List (Delete Cache)</a></li>
+            <li><a href="javascript:;" id="updateLink">Update Youtube-DL</a></li>
         </ul>
+        <ul>
+            <li><a href="?action=refreshCache">Refresh Folder List (Delete Cache)</a></li>
+        </ul>
+        <!--
+        <ul>
+            <li><a href="?refreshCache=true">Use ffmpeg instead of avconv</a></li>
+        </ul>
+        -->
 	</div>
 </div>